All kinds of crazy stuff I did to get a UI to show up
Finally made a fake NetworkInfo 'cause DataConnectionChecker doesn't
work in web pages.
Installed some xcode and cocoapods stuff, deleting rbenv from my box
in the process, in a futile attempt to build an IOs version.
Also, very important and easy to lose sight of with all the other
changes: pulled getting the sharedPreferences instance out of the
registration method so that it happens sooner.